requirements.txt
bson==0.5.8
Click==7.0
colorlog==4.0.2
Flask==1.1.0
Flask-Babel==0.12.2
Flask-DotEnv=0.1.2
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
fuzzywuzzy=0.17.0
httplib2==0.13.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pip=19.1.1
pycodestyle==2.5.0
pymongo==3.8.0
python-dotenv==0.10.3
requests==2.22.0
setuptools==41.0.1
six==1.12.0
Werkzeug==0.15.4
WTForms==2.2.1
